Remove 'assets' dir from history and re-initiate.
Sty's documentation was moved to another repo. Files related to the docs were
removed from the history, except the assets dir.

I decided to move the assets dir to the docs-repo as well.

To keep the source-repo lean, I removed the assets dir from history and
re-initiated it containing only the three images that are used in
README.rst.